Move script to crate root
Move the `script` module out of `primitives` and into the crate root.
Move `opcodes` to the newly created `script` module.

Refactor only, no logic changes.
