+++ bitcoin/src/bip158.rs
@@ -51,8 +51,8 @@ use crate::hashes::{siphash24, Hash};
 use crate::io;
 use crate::prelude::*;
 use crate::primitives::block::Block;
-use crate::primitives::script::Script;
 use crate::primitives::transaction::OutPoint;
+use crate::script::Script;
 use crate::util::endian;
 
 /// Golomb encoding parameter as in BIP-158, see also https://gist.github.com/sipa/576d5f09c3b86c3b1b75598d799fc845
