+++ bitcoin/src/lib.rs
@@ -87,6 +87,7 @@ pub mod error;
 pub mod hash_types;
 pub mod policy;
 pub mod primitives;
+pub mod script;
 pub mod sign_message;
 pub mod util;
 
@@ -101,11 +102,11 @@ pub use crate::consensus::encode::VarInt;
 pub use crate::hash_types::*;
 pub use crate::network::constants::Network;
 pub use crate::primitives::block::{self, Block, BlockHeader, BlockVersion};
+pub use crate::primitives::constants;
 pub use crate::primitives::locktime::{self, absolute, relative};
-pub use crate::primitives::script::{self, Script};
 pub use crate::primitives::transaction::{self, OutPoint, Sequence, Transaction, TxIn, TxOut};
 pub use crate::primitives::witness::{self, Witness};
-pub use crate::primitives::{constants, opcodes};
+pub use crate::script::{opcodes, Script};
 pub use crate::util::amount::{Amount, Denomination, SignedAmount};
 pub use crate::util::ecdsa::{self, EcdsaSig, EcdsaSigError};
 pub use crate::util::key::{KeyPair, PrivateKey, PublicKey, XOnlyPublicKey};
