+++ bitcoin/src/primitives/mod.rs
@@ -9,8 +9,6 @@
 
 pub mod constants;
 pub mod locktime;
-pub mod opcodes;
-pub mod script;
 pub mod transaction;
 pub mod block;
 pub mod witness;
