+++ bitcoin/src/script/opcodes.rs
