+++ bitcoin/src/util/psbt/serialize.rs
@@ -10,7 +10,7 @@ use crate::prelude::*;
 
 use crate::io;
 
-use crate::primitives::script::Script;
+use crate::script::Script;
 use crate::primitives::witness::Witness;
 use crate::primitives::transaction::{Transaction, TxOut};
 use crate::consensus::encode::{self, serialize, Decodable, Encodable, deserialize_partial};
